jm_INTEG/requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
bitsandbytes==0.41.0
blinker==1.6.3
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.7
contourpy==1.1.1
cycler==0.12.1
deepspeed==0.9.5
dlib==19.24.2
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.1.3
face-recognition-models==0.3.0
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.12.4
Flask==3.0.0
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
googletrans==4.0.0rc1
gradio==3.35.2
gradio_client==0.2.9
h11==0.9.0
h2==3.2.0
hjson==3.1.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.18.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
linkify-it-py==2.0.2
lit==17.0.2
# Editable install with no version control (llava==1.1.1)
-e /home/meta-3/mm/llava/LLaVA
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe==2.1.3
matplotlib==3.8.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
ninja==1.11.1.1
numpy==1.26.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
orjson==3.9.9
packaging==23.2
pandas==2.1.1
pathtools==0.1.2
peft==0.4.0
Pillow==10.1.0
protobuf==4.24.4
psutil==5.9.6
py-cpuinfo==9.0.0
pydantic==1.10.13
pydub==0.25.1
Pygments==2.16.1
pypapago==0.1.1.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.10.6
safetensors==0.4.0
scikit-learn==1.2.2
scipy==1.11.3
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.32.0
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
threadpoolctl==3.2.0
timm==0.6.13
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.6
uvicorn==0.23.2
wandb==0.15.12
wavedrom==2.0.3.post3
websockets==11.0.3
Werkzeug==3.0.0
yarl==1.9.2
